moved updateLoggedInState action out of util
